Add native base
Move home screen content to HomeScreen.js
