+++ App.js
@@ -1,22 +1,43 @@
 import "react-native-gesture-handler";
-import React from "react";
+import { StatusBar } from "expo-status-bar";
+import React, { useState, useEffect } from "react";
+import AppLoading from "expo-app-loading";
 import Amplify from "aws-amplify";
 import { withAuthenticator } from "aws-amplify-react-native";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import awsconfig from "./src/aws-exports";
-import HomeScreen from "./src/views/Home";
+import HomeScreen from "./src/views/HomeScreen";
+import * as Font from "expo-font";
+import { Ionicons } from "@expo/vector-icons";
 
 Amplify.configure(awsconfig);
 
 const Stack = createStackNavigator();
 
 const App = () => {
+  const [isReady, setIsReady] = useState(false);
+
+  useEffect(() => {
+    Font.loadAsync({
+      Roboto: require("native-base/Fonts/Roboto.ttf"),
+      Roboto_medium: require("native-base/Fonts/Roboto_medium.ttf"),
+      ...Ionicons.font,
+    });
+
+    setIsReady(true);
+  });
+
+  if (!isReady) {
+    return <AppLoading />;
+  }
+
   return (
     <NavigationContainer>
       <Stack.Navigator>
         <Stack.Screen name="Home" component={HomeScreen} />
       </Stack.Navigator>
+      <StatusBar style="auto" />
     </NavigationContainer>
   );
 };
