+++ package.json
@@ -18,6 +18,9 @@
     "expo": "^40.0.1",
     "expo-status-bar": "~1.0.3",
     "graphql": "^15.5.0",
+    "expo-app-loading": "^1.0.1",
+    "expo-font": "~8.4.0",
+    "native-base": "^2.15.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
